// web-application/src/types-openapi/api/models/CreateFeedbackRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateFeedbackRequest
 */
export interface CreateFeedbackRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateFeedbackRequest
     */
    bookingCourtOccurrenceId: string;
    /**
     * 
     * @type {number}
     * @memberof CreateFeedbackRequest
     */
    rating?: number;
    /**
     * 
     * @type {string}
     * @memberof CreateFeedbackRequest
     */
    comment?: string | null;
    /**
     * 
     * @type {number}
     * @memberof CreateFeedbackRequest
     */
    courtQuality?: number;
    /**
     * 
     * @type {number}
     * @memberof CreateFeedbackRequest
     */
    staffService?: number;
    /**
     * 
     * @type {number}
     * @memberof CreateFeedbackRequest
     */
    cleanliness?: number;
    /**
     * 
     * @type {number}
     * @memberof CreateFeedbackRequest
     */
    lighting?: number;
    /**
     * 
     * @type {number}
     * @memberof CreateFeedbackRequest
     */
    valueForMoney?: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof CreateFeedbackRequest
     */
    mediaUrl?: Array<string> | null;
}

/**
 * Check if a given object implements the CreateFeedbackRequest interface.
 */
export function instanceOfCreateFeedbackRequest(value: object): value is CreateFeedbackRequest {
    if (!('bookingCourtOccurrenceId' in value) || value['bookingCourtOccurrenceId'] === undefined) return false;
    return true;
}

export function CreateFeedbackRequestFromJSON(json: any): CreateFeedbackRequest {
    return CreateFeedbackRequestFromJSONTyped(json, false);
}

export function CreateFeedbackRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateFeedbackRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'bookingCourtOccurrenceId': json['bookingCourtOccurrenceId'],
        'rating': json['rating'] == null ? undefined : json['rating'],
        'comment': json['comment'] == null ? undefined : json['comment'],
        'courtQuality': json['courtQuality'] == null ? undefined : json['courtQuality'],
        'staffService': json['staffService'] == null ? undefined : json['staffService'],
        'cleanliness': json['cleanliness'] == null ? undefined : json['cleanliness'],
        'lighting': json['lighting'] == null ? undefined : json['lighting'],
        'valueForMoney': json['valueForMoney'] == null ? undefined : json['valueForMoney'],
        'mediaUrl': json['mediaUrl'] == null ? undefined : json['mediaUrl'],
    };
}

export function CreateFeedbackRequestToJSON(json: any): CreateFeedbackRequest {
    return CreateFeedbackRequestToJSONTyped(json, false);
}

export function CreateFeedbackRequestToJSONTyped(value?: CreateFeedbackRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'bookingCourtOccurrenceId': value['bookingCourtOccurrenceId'],
        'rating': value['rating'],
        'comment': value['comment'],
        'courtQuality': value['courtQuality'],
        'staffService': value['staffService'],
        'cleanliness': value['cleanliness'],
        'lighting': value['lighting'],
        'valueForMoney': value['valueForMoney'],
        'mediaUrl': value['mediaUrl'],
    };
}