// web-application/src/types-openapi/api/models/DetailFeedbackResponse.ts
export interface DetailFeedbackResponse {
    /**
     * 
     * @type {number}
     * @memberof DetailFeedbackResponse
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof DetailFeedbackResponse
     */
    customerId?: number;
    /**
     * 
     * @type {string}
     * @memberof DetailFeedbackResponse
     */
    bookingCourtOccurrenceId?: string;
    /**
     * 
     * @type {number}
     * @memberof DetailFeedbackResponse
     */
    rating?: number;
    /**
     * 
     * @type {string}
     * @memberof DetailFeedbackResponse
     */
    comment?: string | null;
    /**
     * 
     * @type {number}
     * @memberof DetailFeedbackResponse
     */
    courtQuality?: number;
    /**
     * 
     * @type {number}
     * @memberof DetailFeedbackResponse
     */
    staffService?: number;
    /**
     * 
     * @type {number}
     * @memberof DetailFeedbackResponse
     */
    cleanliness?: number;
    /**
     * 
     * @type {number}
     * @memberof DetailFeedbackResponse
     */
    lighting?: number;
    /**
     * 
     * @type {number}
     * @memberof DetailFeedbackResponse
     */
    valueForMoney?: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof DetailFeedbackResponse
     */
    mediaUrl?: Array<string> | null;
    /**
     * 
     * @type {string}
     * @memberof DetailFeedbackResponse
     */
    adminReply?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof DetailFeedbackResponse
     */
    adminReplyAt?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof DetailFeedbackResponse
     */
    status?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof DetailFeedbackResponse
     */
    createdAt?: Date;
}