// web-application/src/types-openapi/api/models/ListFeedbackRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ListFeedbackRequest
 */
export interface ListFeedbackRequest {
    /**
     * 
     * @type {number}
     * @memberof ListFeedbackRequest
     */
    id?: number | null;
    /**
     * 
     * @type {number}
     * @memberof ListFeedbackRequest
     */
    customerId?: number | null;
    /**
     * 
     * @type {string}
     * @memberof ListFeedbackRequest
     */
    bookingCourtOccurrenceId?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ListFeedbackRequest
     */
    rating?: number | null;
    /**
     * 
     * @type {string}
     * @memberof ListFeedbackRequest
     */
    status?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof ListFeedbackRequest
     */
    from?: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof ListFeedbackRequest
     */
    to?: Date | null;
}

/**
 * Check if a given object implements the ListFeedbackRequest interface.
 */
export function instanceOfListFeedbackRequest(value: object): value is ListFeedbackRequest {
    return true;
}

export function ListFeedbackRequestFromJSON(json: any): ListFeedbackRequest {
    return ListFeedbackRequestFromJSONTyped(json, false);
}

export function ListFeedbackRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListFeedbackRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'customerId': json['customerId'] == null ? undefined : json['customerId'],
        'bookingCourtOccurrenceId': json['bookingCourtOccurrenceId'] == null ? undefined : json['bookingCourtOccurrenceId'],
        'rating': json['rating'] == null ? undefined : json['rating'],
        'status': json['status'] == null ? undefined : json['status'],
        'from': json['from'] == null ? undefined : (new Date(json['from'])),
        'to': json['to'] == null ? undefined : (new Date(json['to'])),
    };
}

export function ListFeedbackRequestToJSON(json: any): ListFeedbackRequest {
    return ListFeedbackRequestToJSONTyped(json, false);
}

export function ListFeedbackRequestToJSONTyped(value?: ListFeedbackRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'customerId': value['customerId'],
        'bookingCourtOccurrenceId': value['bookingCourtOccurrenceId'],
        'rating': value['rating'],
        'status': value['status'],
        'from': value['from'] === null ? null : ((value['from'] as any)?.toISOString()),
        'to': value['to'] === null ? null : ((value['to'] as any)?.toISOString()),
    };
}