// web-application/src/types-openapi/api/models/ListFeedbackResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ListFeedbackResponse
 */
export interface ListFeedbackResponse {
    /**
     * 
     * @type {number}
     * @memberof ListFeedbackResponse
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof ListFeedbackResponse
     */
    customerId?: number;
    /**
     * 
     * @type {string}
     * @memberof ListFeedbackResponse
     */
    bookingCourtOccurrenceId?: string;
    /**
     * 
     * @type {number}
     * @memberof ListFeedbackResponse
     */
    rating?: number;
    /**
     * 
     * @type {string}
     * @memberof ListFeedbackResponse
     */
    comment?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListFeedbackResponse
     */
    status?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof ListFeedbackResponse
     */
    createdAt?: Date;
}

/**
 * Check if a given object implements the ListFeedbackResponse interface.
 */
export function instanceOfListFeedbackResponse(value: object): value is ListFeedbackResponse {
    return true;
}

export function ListFeedbackResponseFromJSON(json: any): ListFeedbackResponse {
    return ListFeedbackResponseFromJSONTyped(json, false);
}

export function ListFeedbackResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListFeedbackResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'customerId': json['customerId'] == null ? undefined : json['customerId'],
        'bookingCourtOccurrenceId': json['bookingCourtOccurrenceId'] == null ? undefined : json['bookingCourtOccurrenceId'],
        'rating': json['rating'] == null ? undefined : json['rating'],
        'comment': json['comment'] == null ? undefined : json['comment'],
        'status': json['status'] == null ? undefined : json['status'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
    };
}

export function ListFeedbackResponseToJSON(json: any): ListFeedbackResponse {
    return ListFeedbackResponseToJSONTyped(json, false);
}

export function ListFeedbackResponseToJSONTyped(value?: ListFeedbackResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'customerId': value['customerId'],
        'bookingCourtOccurrenceId': value['bookingCourtOccurrenceId'],
        'rating': value['rating'],
        'comment': value['comment'],
        'status': value['status'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
    };
}